ms-facturacion/requirements.txt
flask
pika
mysql-connector-python
requests
